fix: make khelm fails on osx due to leading spaces in helm version.
When go building the khelm binary, an ldflags string is concatenated and passed
as an argument to go build. Part of this string is a helm version, which is
extracted from go.mod. The extracted string contains a number of leading spaces,
which subsequently causes the go build to fail. The malformed command previously
looked something like:

    go build … -ldflags "… -X main.helmVersion=  v2.17.0"
                                               ^^- extra spaces